Extracting Line Chart into separate component with it's own data state
